Added accompanist libraries for navigation and changed removed specific display style menu for explore screen
