+++ gradle/libs.versions.toml
@@ -1,4 +1,5 @@
 [versions]
+accompanist-version = "0.30.0"
 activity-compose-version = "1.7.0"
 android-version = "8.1.0-alpha11"
 android-junit-version = "1.1.5"
@@ -30,6 +31,10 @@ compose-compiler-version = "1.4.3"
 
 
 [libraries]
+#   Accomanist
+accompanist-animation-navigation = { module = "com.google.accompanist:accompanist-navigation-animation", version.ref = "accompanist-version" }
+accompanist-bottomsheet-navigation = { module = "com.google.accompanist:accompanist-navigation-material", version.ref = "accompanist-version" }
+accompanist-system-ui-controller = { module = "com.google.accompanist:accompanist-systemuicontroller", version.ref = "accompanist-version" }
 #   Compose Activity
 activity-compose = { module = "androidx.activity:activity-compose", version.ref = "activity-compose-version" }
 #   Kotlin KTX
@@ -75,8 +80,6 @@ room-ktx = { module = "androidx.room:room-ktx", version.ref = "room-version" }
 room-paging = { module = "androidx.room:room-paging", version.ref = "room-version" }
 #   Shimmer
 shimmer-effect = { module = "com.valentinilk.shimmer:compose-shimmer", version.ref = "shimmer-version" }
-#   System UI Controller
-system-ui-controller = { module = "com.google.accompanist:accompanist-systemuicontroller", version.ref = "system-ui-controller-version" }
 #   Timber (Logging)
 timber = { module = "com.jakewharton.timber:timber", version.ref = "timber-version" }
 #   Testing
@@ -115,9 +118,7 @@ android-application-plugin = { id = "com.android.application", version.ref = "an
 android-library-plugin = { id = "com.android.library", version.ref = "android-version" }
 kotlin-android-plugin = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin-android-version" }
 
-
 dagger-hilt-plugin = { id = "com.google.dagger:hilt-android-gradle-plugin", version.ref = "dagger-hilt-version" }
-ksp = { id = "com.google.devtools.ksp", version.ref = "ksp-version" }
 
 kotlin-kapt-plugin = { id = "org.jetbrains.kotlin.kapt", version.ref = "kotlin-kapt-version" }
 dagger-hilt-module-plugin = { id = "com.google.dagger.hilt.android", version.ref = "dagger-hilt-version" }
